# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: youtube
    description: YouTube download operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /youtube/info:
    get:
      operationId: getVideoInfo
      tags: [youtube]
      summary: Get YouTube video info
      description: Fetches metadata and available formats for a YouTube video
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
          description: YouTube video URL
      responses:
        "200":
          description: Video info with available formats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoInfo"
        "400":
          description: Invalid URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to fetch video info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    VideoFormat:
      type: object
      properties:
        formatId:
          type: string
        quality:
          type: string
        ext:
          type: string
        resolution:
          type: ["string", "null"]
        filesize:
          type: ["integer", "null"]
        fps:
          type: ["number", "null"]
        vcodec:
          type: ["string", "null"]
        acodec:
          type: ["string", "null"]
        hasVideo:
          type: boolean
        hasAudio:
          type: boolean
      required:
        - formatId
        - quality
        - ext
        - resolution
        - filesize
        - fps
        - vcodec
        - acodec
        - hasVideo
        - hasAudio
    VideoInfo:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        thumbnail:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        uploader:
          type: ["string", "null"]
        viewCount:
          type: ["integer", "null"]
        formats:
          type: array
          items:
            $ref: "#/components/schemas/VideoFormat"
      required:
        - id
        - title
        - thumbnail
        - duration
        - uploader
        - viewCount
        - formats